using_lightning/pyproject.toml
[tool.poetry]
name = "blog-lightning"
version = "0.1.0"
description = "A sample project for using PyTorch Lightning"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.3.2"
lightning = "^2.2.0"
torchvision = "^0.17.0"
torchmetrics = "^1.3.0.post0"
scipy = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
